moved db functions to separate file
